Revise fusion pipeline script to use parameter enum
